Hexadecimal : Change from int to long

+++ 0x01-variables_if_else_while/8-print_base16.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	int ch, hex;
+	long ch, hex;
 
 	for (ch = '0'; ch < '16'; ch++)
 	{
